requirements.txt
google-genai
google-generativeai
tavily-python
requests
pillow
moviepy
ffmpeg-python
python-dotenv
beautifulsoup4
html2image
